requirements.txt
asgiref==3.8.1
dj-database-url==3.0.1
Django==5.2.3
django-cotton==2.1.2
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.0
uv==0.7.17
